+++ src/Controllers/Dashboard/Dashboard.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace lovis.Controllers.Dashboard
+{
+    public class Dashboard
+    {
+    }
+}