+++ src/Controllers/Proyects/Templates/Templates.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace lovis.Controllers.Proyects.Templates
+{
+    public class Templates
+    {
+        #region ATTRIBUTES
+
+        public int Id { get; set; }
+        public string Title { get; set; }
+        public string Type1 { get; set; }
+        public string Type2 { get; set; }
+        public string Type3 { get; set; }
+        public string Type4 { get; set; }
+        public string Type5 { get; set; }
+
+        #endregion
+
+        #region CONSTRUCTORS
+
+        #endregion
+
+        #region NEW
+
+        #endregion
+
+        #region EDIT
+
+        #endregion
+
+        #region REMOVE
+
+        #endregion
+    }
+}