+++ src/Controllers/UserLicense/UserLicense.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace lovis.Controllers.UserLicense
+{
+    public class UserLicense
+    {
+        #region ATTRIBUTES
+
+        public int Id { get; set; }
+        public string IdUser { get; set; }
+        public string IdLicense { get; set; }
+        public int Role { get; set; }
+        public bool isBlocked { get; set; }
+
+        #endregion
+
+        #region CONSTRUCTORS
+
+        #endregion
+
+        #region NEW 
+
+        #endregion
+
+        #region EDIT 
+
+        #endregion
+    }
+}