+++ src/Controllers/Users/Register/Register.cs
@@ -7,5 +7,24 @@ namespace lovis.Controllers.Users.Register
 {
     public class Register
     {
+        #region ATTRIBUTES
+
+        #endregion
+
+        #region CONSTRUCTORS
+
+        #endregion
+
+        #region NEW
+
+        #endregion
+
+        #region EDIT
+
+        #endregion
+
+        #region REMOVE
+
+        #endregion
     }
 }