+++ src/Controllers/Users/Users.cs
@@ -7,5 +7,36 @@ namespace lovis.Controllers.Users
 {
     public class Users
     {
+        #region ATTRIBUTES
+
+        public string Id { get; set; }
+        public string Username { get; set; }
+        public string RandomKey { get; set; }
+        public string FinishedKey { get; set; }
+        public string Name { get; set; }
+        public string Surname { get; set; }
+        public string Phone1 { get; set; }
+        public string Phone2 { get; set; }
+        public string Address { get; set; }
+        public DateTime DateRegister { get; set; }
+        public string idLicense { get; set; }
+
+        #endregion
+
+        #region CONSTRUCTORS
+
+        #endregion
+
+        #region NEW
+
+        #endregion
+
+        #region EDIT
+
+        #endregion
+
+        #region REMOVE
+
+        #endregion
     }
 }